refactor: import shared types from formResultTypes directly
Drop the backward-compat re-exports from FormResult.ts and point the
two internal consumers (FormResult.test, FormModal) at formResultTypes
where the types now live, so there is one canonical import path per
symbol.

Addresses review feedback on PR #482.
