+++ src/FormModal.ts
@@ -2,7 +2,8 @@ import { throttle } from "@std";
 import { App, Modal, Setting } from "obsidian";
 import { SvelteComponent } from "svelte";
 import FormModalComponent from "./FormModal.svelte";
-import FormResult, { type ModalFormData } from "./core/FormResult";
+import FormResult from "./core/FormResult";
+import type { ModalFormData } from "./core/formResultTypes";
 import { formDataFromFormDefaults } from "./core/formDataFromFormDefaults";
 import type { FormDefinition, FormOptions } from "./core/formDefinition";
 import { FormEngine, makeFormEngine } from "./store/formEngine";
